Make serialized data types consistent with defined data types

+++ src/nupic/proto/BitHistory.capnp
@@ -7,12 +7,12 @@ struct BitHistoryProto {
   stats @1 :List(Stat);
   lastTotalUpdate @2 :UInt32;
   learnIteration @3 :UInt32;
-  alpha @4 :Float32;
+  alpha @4 :Float64;
   verbosity @5 :UInt8;
 
   # Next ID: 2
   struct Stat {
     index @0 :UInt32; # Check if actually unsigned
-    dutyCycle @1 :Float32;
+    dutyCycle @1 :Float64;
   }
 }
